Extract imbi.endpoints.components.models
